Test unicode values in form fields (currently failing)
This test was inspired by failures in the tests of a real
world application which uses zope.testbrowser to test that
forms handle unicode.

The test assumes that "sane" means "compatible with 
zope.testbrowser 4 on python 2". But that might be wildly
off-base. Not sure what the intention of the python 3
port wrt unicode is and if that is meant to change how
things worked under python 2.