Fix tests on master
browser.cookies.change() helpfully makes the cookie disappear if you set
an expirationd date in the past.

We had a narrative test that tried to use a fixed future date (2019
January 1st) that is no longer a future date.

Now in a regular unit test I could do things like datetime.date.today()
and write a test that will never expire.  But this is a doctest so I'll
just shift all the dates 10 years into the future and leave the problem
for some unlucky future person down the road.

Fixes #62.