Make the migration example more explicit
I was looking at the migration example example the other day and
wondered where to specify the filename of my ftesting.zcml.  The old
code had it in the open, the new one has it in a default __init__
argument in one of the two base classes, not exactly trivial to find.

I think this can be useful to other people who migrate from old
zope.testbrowser too.