Stop sending Referer on browser.open/browser.post
RFC 2616 14.36 says: "The Referer field MUST NOT be sent if the
Request-URI was obtained from a source that does not have its own URI,
such as input from the user keyboard."  In the context of a test
browser, `browser.open` is analogous to input from the user keyboard and
should not result in sending a `Referer` header.

We save the previous request's referrer in order to maintain the
appropriate state for `browser.reload`.

Fixes #87.